refactor: 누락된 dto 패키지 추가
