+++ backend/src/main/java/com/staccato/auth/controller/AuthController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.staccato.auth.service.AuthService;
-import com.staccato.auth.service.request.LoginRequest;
-import com.staccato.auth.service.response.LoginResponse;
+import com.staccato.auth.service.dto.request.LoginRequest;
+import com.staccato.auth.service.dto.response.LoginResponse;
 
 import lombok.RequiredArgsConstructor;
