+++ backend/src/main/java/com/staccato/auth/service/dto/response/LoginResponse.java
@@ -1,0 +1,4 @@
+package com.staccato.auth.service.dto.response;
+
+public record LoginResponse(String token) {
+}
