--- backend/src/main/java/com/staccato/auth/service/response/LoginResponse.java
@@ -1,4 +1,0 @@
-package com.staccato.auth.service.response;
-
-public record LoginResponse(String token) {
-}
